pyproject.toml
[tool.poetry]
name = "riptide"
version = "0.4.0"
description = "Bifrost Object Detection Evaluation Suite"
authors = ["huangjoel <joel@bifrost.ai>", "Sean <sean_yap@bifrost.ai"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.19"
torch = "^1.12"
pandas = "^1.4"
torchvision = "^0.13.0"
jinja2 = "^3.1.2"
termcolor = "^2.0.1"
ujson = "^5.5.0"
matplotlib = "^3.6.1"
tqdm = "^4.65.0"
plotly = "^5.14.0"
scikit-learn = "^1.2.2"
pydantic = "^1.10.7"
hdbscan = "^0.8.29"
shapely = "^2.1.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
isort = "^5.10.1"
black = "^22.3.0"
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
hypothesis = {extras = ["numpy"], version = "^6.75.3"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
preview = true

[tool.isort]
profile = "black"
